+++ inc/field/datefield.class.php
@@ -42,7 +42,7 @@ class DateField extends DatetimeField
 
    public function getRenderedHtml($domain, $canEdit = true): string {
       if (!$canEdit) {
-         return $this->value;
+         return \Glpi\Toolbox\Sanitizer::encodeHtmlSpecialChars($this->value);
       }
 
       $html = '';