+++ inc/field/textfield.class.php
@@ -94,9 +94,17 @@ class TextField extends PluginFormcreatorAbstractField
    }
 
    public function deserializeValue($value) {
-      $this->value = ($value !== null && $value !== '')
-         ? $value
-         : '';
+      $isUrlPrefilled = isset($_SESSION['formcreator']['url_prefilled']) && $_SESSION['formcreator']['url_prefilled'];
+      
+      if ($value !== null && $value !== '') {
+         if (!$isUrlPrefilled) {
+            $this->value = \Glpi\Toolbox\Sanitizer::encodeHtmlSpecialChars($value);
+         } else {
+            $this->value = $value;
+         }
+      } else {
+         $this->value = '';
+      }
    }
 
    public function getValueForDesign(): string {